+++ client/main.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { render } from 'react-dom';
 import { Sites } from '../imports/api/sites';
-import App from './App';
+
+import App from '../imports/ui/App';
  
 Meteor.startup(() => {
   render(<App />, document.getElementById('render-target'));
