+++ imports/ui/components/sites/List.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { Sites } from '../../../imports/api/sites';
+import { Sites } from '../../../api/sites';
 
 class Cells extends React.Component {
     render() {
